requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
reportlab>=4.0.0
langdetect>=1.0.9
sentence-transformers>=2.2.2
textstat>=0.7.3